Rewrite proxy setup to include Terraform in image
This change updates `proxy/README.md` to reflect how the Docker image can be
used as a base image. This README will be used as the description on Docker
Hub.

With this change the image now includes Terraform but allows extending at build
time. Terraform isn't small but it works.