Combine region_map_keys and region_map_values into region_map
This change combines the `region_map_keys` and `region_map_values` into a single
type that has the keys and values together while ALSO preserving the order of
the keys. The order is important for setting the fallback pool as well as to indicate
the failover priority order (`cloudflare_load_balancer.default_pool_ids`).

Note: the built-in map type does not have a way to retrieve keys while
preserving the order. As we need the first pool to use it as the fallback
we cannot simply use a map.